pyproject.toml
[tool.poetry]
name = "projections"
version = "0.1.0"
description = ""
authors = ["Luis Da Silva <luisdanield@gmail.com>"]
packages = [
    { include = "projections", from = "src"}
]

[tool.poetry.dependencies]
python = "^3.8"
pyshp = "^2.1.3"
tqdm = "^4.60.0"
pandas = "^1.2.4"
numpy = "^1.20.3"
Pillow = "^8.2.0"
PyYAML = "^5.4.1"
geopandas = "^0.9.0"
Shapely = "^1.7.1"
pyarrow = "^4.0.1"
pygeos = "^0.10"
matplotlib = "^3.4.2"
rasterio = "^1.2.6"
rioxarray = "^0.4.3"
pytest = "^6.2.4"
netCDF4 = "^1.5.7"
h5py = "^3.4.0"
tables = "^3.6.1"
pyhdf = "^0.10.3"
sklearn = "^0.0"

[tool.poetry.dev-dependencies]
jupyter = "^1.0.0"
black = {version = "^21.6b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
